requirements.txt
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
firecrawl-py==0.0.16
idna==3.10
nest-asyncio==1.6.0
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.1
